revert r33387. the emulation breaks (eg trap17.prg) when reg_pc is uint16_t. this is really ugly and needs to be fixed.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@34095 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/mainc64cpu.c
@@ -534,7 +534,8 @@ inline static int interrupt_check_irq_delay(interrupt_cpu_status_t *cs,
 /* ------------------------------------------------------------------------- */
 
 #ifdef NEED_REG_PC
-uint16_t reg_pc;
+/* FIXME: this should really be uint16_t, but it breaks things (eg trap17.prg) */
+unsigned int reg_pc;
 #endif
 
 static uint8_t **o_bank_base;
@@ -560,7 +561,8 @@ void maincpu_mainloop(void)
     uint8_t flag_n = 0;
     uint8_t flag_z = 0;
 #ifndef NEED_REG_PC
-    uint16_t reg_pc;
+    /* FIXME: this should really be uint16_t, but it breaks things (eg trap17.prg) */
+    unsigned int reg_pc;
 #endif
     uint8_t *bank_base;
     int bank_start = 0;
