+++ vice/src/maincpu.c
@@ -387,7 +387,8 @@ inline static int interrupt_check_irq_delay(interrupt_cpu_status_t *cs,
 /* ------------------------------------------------------------------------- */
 
 #ifdef NEED_REG_PC
-uint16_t reg_pc;
+/* FIXME: this should really be uint16_t, but it breaks things (eg trap17.prg) */
+unsigned int reg_pc;
 #endif
 
 static uint8_t **o_bank_base;
